requirements.txt
-e .
alabaster==0.7.10
alembic==0.9.6
apipkg==1.4
astroid==1.5.3
Babel==2.5.1
bandit==1.4.0
beautifulsoup4==4.6.0
bleach==2.1.2
blinker==1.4
bson
certifi==2017.11.5
chardet==3.0.4
click==6.7
codacy-coverage==1.3.7
coverage==4.4.2
cycler==0.10.0
decorator==4.1.2
docutils==0.14
dominate==2.3.1
entrypoints==0.2.3
execnet==1.5.0
flake8==3.5.0
flake8-quotes==0.12.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-Login==0.4.1
flask-nav==0.6
Flask-Session==0.3.1
Flask-WebTest==0.0.9
Flask-WTF==0.14.2
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
isort==4.2.15
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.4.0
jupyterlab==0.30.6
jupyterlab-launcher==0.6.0
lazy-object-proxy==1.3.1
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pyparsing==2.2.0
pytest==3.2.3
pytest-cache==1.0
pytest-cov==2.5.1
pytest-mccabe==0.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
requests==2.18.4
rope==0.10.7
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.15
stevedore==1.27.1
terminado==0.8.1
testfixtures==5.3.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
visitor==0.1.3
waitress==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.7.4
WebTest==2.0.29
Werkzeug==0.12.2
widgetsnbextension==3.0.8
wrapt==1.10.11
WSGIProxy2==0.4.4
WTForms==2.1
